Added CharacterController - KnownCharacters doesnt work yet

+++ Rawdata_Porfolio_2/Entity Framework/Domain/title.cs
@@ -9,12 +9,12 @@ namespace Rawdata_Porfolio_2.Pages.Entity_Framework.Domain
     public class Title
     {
         [Key]
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public string Type { get; set; }
         public bool IsAdult { get; set; }
-        public int? Year_Start { get; set; }
-        public int? Year_End { get; set; }
-        public int? Runtime { get; set; }
+        public Int16 ? Year_Start { get; set; }
+        public Int16 ? Year_End { get; set; }
+        public Int32 ? Runtime { get; set; }
         public double ? AvgRating { get; set; }
         public string Poster { get; set; }
         public string Plot { get; set; }
